backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
pydicom>=2.4
numpy>=1.24
dipy>=1.7
nibabel>=5.1
scipy>=1.10
scikit-image>=0.21
django-cors-headers>=4.3.1
matplotlib>=3.7
